separet getDebugTree file+test
